+++ CHANGELOG.md
@@ -20,11 +20,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- [!370][!370] Increased the precedence of dereference and bitwise operators
 
 ### Removed
 
 <!--Links:-->
 [!365]: https://gitlab.com/spade-lang/spade/-/merge_requests/365
+[!370]: https://gitlab.com/spade-lang/spade/-/merge_requests/370
 
 ## [0.11.0] - 2024-11-28